Trigger the accordion when the topic is collapsed
We are now allowing the content designers in the publisher to decide whether the accordion is applied or not. There wasn't a clear set of rules that we could define. It seems as though a human needs to decide whether it is collapsed or not.

The default is that the topic sections are expanded. It is more likely that a topic will be collapsed but the downside of a difficult to use topic is more significant than the upside of saving a click in the publisher.