Save the state of the accordion using session storage
When a section is opened - save its ID to local storage and when it is
closed, use this ID as a key to remove the item.

When the page loads, check session storage to see if any items have
been stored and if so, open that section and set the aria-expanded
attribute of the button true.