Save opened sections to session storage
Create a prefix for each service manual section to be saved, so each
topic page is unique when saved to session storage. This assumes each
topic has a different title and each section has a unique title.

When the page loads, check for saved sections in local storage with a
key which matches “GOVUK_service_manual_” and the topic name. If it
matches, then use the ID of the saved section to open that section.

Also update the aria-expanded attribute of the button and check to see
if the open/close all text must be updated. Change this to “close all”
if all sections are opened.