Add js hooks for StickAtTopWhenScrolling and HighlightActiveSection
The navigation highlight module uses this hook for the
in-page navigation (js-page-contents).

Also add a js hook to the page contents area
This is required to get the stick-at-top-when-scrolling js to work.

Stick at top when scrolling also has a futher optional hook to set the
sticky element and the shim to resize to fit the parent (the hook for this is js-sticky-resize).