Service standard points are now children in links
It turns out it's possible to patch links with a 'parent' link_type on a child and the the child will be expanded in the parent's links. This makes for a simpler operation in the publisher.

Here we need to pull the points from the 'children' array rather than the custom 'points' array we had previously created.