Guide category
A point page does not have a topic but we do want to link to the service standard in the breadcrumb. We rename the idea of a topic to a category, that can either be a topic, if it exists, or the parent. A normal guide will have a topic and a point will have a parent.

It can handle having neither.