Set up Wraith correctly for service manual publisher
Change the base URL to use the hostname exposed by vagrant dns. Taking screenshots does not work from within the VM because assets-origin.dev.gov.uk is not resolvable within the VM so none of the assets load.

Change the paths to point to actual paths within the service manual. The existing paths had not changed since the repository was forked from government-frontend, and made no sense for this application.