Don't render blank headings for topics sections without titles
Some topics have been configured with only one section which has neither a name nor a description.

Previously the front-end did not handle that case very well – it still rendered a heading for the section, including a heading and description. This lead to blank h2s which is bad from an accessibility point of view (and also just not a great thing to do)

Now, for topic sections without a title we omit the title and description. Additionally, if the section has a title but not a description, we omit the markup for the description.