Defend against public_updated_at not being available
Before a content item is published it does not have the `public_updated_at` timestamp added by the publishing-api because it has not yet been published. This was causing an error when previewing the item on the draft stack.

We add `#visible_updated_at` which falls back on the `updated_at` timestamp when we're previewing the first draft.