Improve this page
A widget to offer a suggestion on how to improve the page or to feedback that the page was useful. This is an alternative to the existing 'Report a Problem' widget that gives encourages positive feedback.

It works by submitting an AJAX request to the feedback frontend, which in turn submits a request to the support-api, which in turn submits the feedback to Zendesk.

Custom event tracking via Google Analytics is supported if the GOV.UK Analytics module is loaded.

The fallback behaviour in the absense of javascript is to send the user to the GOV.UK contact page.