Screen reader compatibility for Improve this page fields
A hidden 'required' has been placed in the "How should we improve this page?" label. This should hopefully make it clearer that the field is required. It is much clearer to a user with unimpared vision that the field is required because the nearby name and email fields are optional, thereby implying that the description field isn't optional.