Remove bottom padding from Improve this page prompt
The prompt is a blue box on top of a blue border. This makes the text in the prompt seem as though it isn't vertically aligned properly as there is more blue below the text than above it. This removes the padding at the bottom so the effect is diminished.