Don’t use HTML5 browser ‘required’ validation
The HTML5 validation is provided by the browser and we are not able to style the error messages. This means that the validation experience is inconsistent with other parts of GOV.UK and with any more advanced validation that might happen server side.

There is already server side validation for the description being required, so the simplest fix is just to remove the required attribute and fall back to that.