Hide "improve this page" bar and form when printing
At the moment, when printing a page from the service manual that has the
"Is this page useful?" bar at the bottom, it prints the feedback form.

As a printed out feedback form is by and large not particularly useful
to our users, this adds a "print" stylesheet which hides the
`improve-this-page` container.