Refactor header search logic
We do not wish to display the search box on the service toolkit page, like we currently exclude it on the homepage. This replaces the too-specific `is_homepage?` method (which was only used for this purpose) with a new `include_search_in_header?` method and updates the content items controller to use that instead when configuring search in slimmer.