Exclude footer links from footer
The links that appear in the top half of the footer (.footer-categories) replicate the home page of GOV.UK and represent the needs of the citizens using the site.

As the primary users of the Service Manual (and forthcoming Service Toolkit and Product pages) are people within government building a services these links are not relevant and may even compete (could be confused with Service Manual content).