Full width heros on toolkit and manual homepage
This follows a similar pattern to the GOV.UK homepage, overriding #wrapper when the body has a `full-width` class to stop it being a site width container, and adding the site width container within the hero and the body of the page.

We also need to wrap the beta banner and feedback component so that they’re centered.