Remove old jenkins.sh, jenkins-schema.sh files
The jenkins.sh file has been replaced by a Jenkinsfile which was added in the previous commit.

Although we have a jenkins-schema.sh file, this only seems to exist because it was present when this repository was forked from government-fronted. We don’t actually have any existing schemas jobs set in Jenkins, nor do we have any status checks enabled for the schemas repo.

Therefore, we can remove both files for now and we can hook up schema tests later, should we decide that would actually be useful.