Fix padding on page contents
This prevents the page navigation from ‘jumping’ back down when it reaches the top of the page – by using padding rather than margin, it’ll stick as soon as the padded element hits the top of the screen.

We still need to use margin on mobile as we’re relying on margins collapsing for the spacing before the page contents.