Fix JavaScript tests of accordion session storage
The Jasmine tests for the accordion state were manipulating the state
themselves, which meant that they were not fully testing the application
code.

This fixes the problem by removing all state changes in the tests,
except for resetting the session storage after each test.

Some jQuery selectors have also been updated from `$('foo')` to
`$element.find('foo')` so that the selectors search the accordion
element rather than the whole document, which is not set up by the test.