Raise better errors for unsupported formats
Users will often visit the root of an application when they first start it, to make sure it is working. At the minute, we throw an exception if you try to do this. This is intentional – applications should serve on the routes as they would appear on GOV.UK, so the service manual shouldn’t do anything for `/` – it’s homepage is at `/service-manual`.

However, the exception we throw (No support for format special_route) does not make this obvious. This handles the root as a special case, tries to makes the general error message more useful as well, and adds some documentation to the README.