Fix broken accordion in IE7 / IE11 compat. mode
Accessing individual characters in strings as if they are arrays is an ECMAScript 5 feature which was only introduced in IE8. So IE 7 and below (and IE11 when in compatibility mode) are erroring because of this.

Notably this is leaving accordions in a state where the original page content has been removed, but the accordion has not been initialised, leaving the user with no way to navigate further.