Allow for modules to not be initialised
Because we’re now loading modules.js from frontend toolkit, it doesn’t get loaded when tests are run, so we need to guard against that. We don’t actually need the logic in modules.js as the tests call start directly on the module, we just need window.GOVUK.Modules to exist.