Fix alignment of feedback form on Service Toolkit
For the service toolkit, we override the wrapper to make it full width to allow for the full width masthead. This means that it doesn’t center the feedback form and its toggle as it does on other pages on GOV.UK.

Sadly this means we need to add more hacks to explicitly center the feedback form and its toggle when this full width override is in place.