Don't send User Agent in contact form
The "How should we improve this page?" currently sends the user agent
of the request.

Like https://github.com/alphagov/govuk_publishing_components/pull/699,
this is wrong because the page will be cached, so that any subsequent
page requests will return the same HTML with the same form fields - and
the previous request's user agent.

Because this feature has been broken for a number of years, we're
removing it instead of fixing it. We think the reason the user agent
was included in the first place was taken from GOV.UK. We've found no
good reason to collect this information.