Repoint email signups to generic API
Previously we relied on each topic (and the service standard) having
an associated '/email-signup' content item. The matching criteria are
the topic itself, with a restriction on the document type [1]. The
additional document type restriction turns out to be redundant, since
service manual guides are the only document type linked to service
manual topics (and similarly for the service standard). This repoints
the email signup links to the generic page on Email Alert Frontend,
which supports signups to a given content item.

[1]: https://github.com/alphagov/service-manual-publisher/blob/master/app/presenters/topic_email_alert_signup_presenter.rb#L53