Remove poltergeist
The poltergeist gem was added in #10 to test the accordion but the
associated tests were removed in #581 when it was changed to use
GOV.UK Frontend v3. This also allows us to remove the Action Controller
monkey patch since the Slimmer test helper handles the requests for
the static templates.