Pin rails to specific version and make it stand out
The rails gem is the root of our dependency graph and should be used to
drive the versions of other dependencies. If we don't specify the
version we can get stuck we try to update rails independently, e.g.

  % bundle update rails
  Fetching gem metadata from https://rubygems.org/..........
  Fetching gem metadata from https://rubygems.org/.
  Resolving dependencies......
  ...
  Bundler attempted to update rails but its version stayed the same
  Bundle updated!

However if we specify the specific version of rails we're targeting
then bundler can provide us with hints that allow us to identify the
other gems that need updating, e.g.

  % bundle update rails
  Fetching gem metadata from https://rubygems.org/..........
  Fetching gem metadata from https://rubygems.org/.
  Resolving dependencies...
  Bundler could not find compatible versions for gem "railties":
    In Gemfile:
      rails (= 6.0.3.2) was resolved to 6.0.3.2, which depends on
        railties (= 6.0.3.2)

      rails_translation_manager was resolved to 0.1.0, which depends on
        rails-i18n was resolved to 5.1.3, which depends on
          railties (>= 5.0, < 6)

      sass-rails was resolved to 5.0.7, which depends on
        railties (>= 4.0.0, < 6)

These hints allow us to run the following command:

  % bundle update rails rails-i18n sass-rails
  Fetching gem metadata from https://rubygems.org/..........
  Fetching gem metadata from https://rubygems.org/.
  Resolving dependencies....
  ...
  Bundle updated!

which resolves correctly without having to do a full `bundle update`.
A full bundle update results in a large amount amount of dependency
changes and is generally not a good idea as it makes it hard to track
what change introduced a problem.

The `gem "rails", "6.0.3.2"` line is moved to the top of the Gemfile
to signify its importance and to separate it out from the rest of the
entries which won't have constraints unless required for reasons that
should be documented by inline comments on the relevant entries.